Refactors integration tests to be more isolated
This makes all integration tests use unique data. Notably, they no
longer use constant trace ID or service names, as this can lead to
difficulty when a keyspace is re-used. For example, service names and
trace IDs are often partitions. When debugging it is easier when test
data is isolated by a service name as it can be easily queried.

This also pulls the heaviest tests into their own classes so that they
don't overload the storage containers used by the bulk of our tests.

Finally, this fixes a few glitches in the v1 cassandra storage.